api/requirements.txt
streamlit
fastapi
uvicorn
tensorflow
pillow
